requirements.txt
configparser
flask
flask_sqlalchemy
Flask-WTF
googlemaps~=4.5.3
pyyaml
requests
SQLAlchemy
psycopg2
oauth2client
setuptools
psycopg[binary]
python-dotenv